Move to es2017 target, simplify tsconfig
