+++ tests/tsconfig.json
@@ -1,22 +1,10 @@
 {
+  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop": true,
-    "jsx": "react",
-    "lib": ["dom", "es2015"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noImplicitAny": true,
-    "noUnusedLocals": true,
     "outDir": "build",
-    "preserveWatchOutput": true,
-    "resolveJsonModule": true,
-    "rootDir": "..",
-    "skipLibCheck": true,
-    "strict": true,
-    "strictNullChecks": false,
-    "target": "es2015",
+    "rootDir": "src",
+    "target": "es2017",
     "types": ["jest"]
-  }
+  },
+  "include": ["src/*"]
 }
