+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
+    "composite": true,
     "declaration": true,
     "jsx": "react",
-    "lib": ["es2015", "dom"],
-    "module": "commonjs",
+    "module": "esnext",
     "moduleResolution": "node",
     "noEmitOnError": true,
     "noImplicitAny": true,
@@ -13,7 +13,7 @@
     "preserveWatchOutput": true,
     "resolveJsonModule": true,
     "rootDir": "src",
-    "target": "es2015",
+    "target": "es2017",
     "strict": true,
     "strictNullChecks": false,
     "types": []
